Add a windows.rs file for terminal
