+++ crates/sshx/Cargo.toml
@@ -14,11 +14,10 @@ aes = "0.8.3"
 ansi_term = "0.12.1"
 anyhow.workspace = true
 argon2 = { version = "0.5.2", default-features = false, features = ["alloc"] }
+cfg-if = "1.0.0"
 clap.workspace = true
-close_fds = "0.3.2"
 ctr = "0.9.2"
 encoding_rs = "0.8.31"
-nix = { version = "0.27.1", features = ["ioctl", "process", "signal", "term"] }
 pin-project = "1.1.3"
 sshx-core.workspace = true
 tokio.workspace = true
@@ -27,3 +26,10 @@ tonic.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 whoami = { version = "1.5.1", default-features = false }
+
+[target.'cfg(unix)'.dependencies]
+close_fds = "0.3.2"
+nix = { version = "0.27.1", features = ["ioctl", "process", "signal", "term"] }
+
+[target.'cfg(windows)'.dependencies]
+conpty = "0.5.1"
